feat(commands): get single rule
